requirements.txt
boto3==1.4.5
flask==0.12.2
gunicorn==19.7.1
pyserial==3.4
pyyaml==3.12